Navigation added

+++ frontend/src/components/admin/ArtistList.js
@@ -14,11 +14,13 @@ const ArtistList = () => {
     const alert = useAlert()
     const dispatch = useDispatch()
     const navigate = useNavigate()
-    const { loading, error, artists } = useSelector(state => state.artists)
-    const { error: deleteError, isDeleted } = useSelector(state => state.artist)
+    const { loading, error, artists       } = useSelector( state => state.artists )
+    const { error: deleteError, isDeleted } = useSelector( state => state.artist )
 
     useEffect(() => {
+
         dispatch(getArtists())
+
         if(error) {
             alert.error(error)
             dispatch(clearErrors())
@@ -28,14 +30,17 @@ const ArtistList = () => {
             dispatch(clearErrors())
         }      
         if(isDeleted) {
-            alert.success('Artist deleted successfully')            
+            alert.success('Artist Deleted Successfully')            
             dispatch({ type: DELETE_ARTIST_RESET })
             navigate('/admin/artists')
         }
+
     }, [dispatch, navigate, alert, error, isDeleted, deleteError])
 
     const deleteCategoryHandler = (id) => {
-        dispatch(deleteArtist(id))
+        if ( window.confirm("Are you Sure?") === true ) {
+            dispatch(deleteArtist(id))
+        }         
     }
 
     const setCategories = () => {
@@ -44,7 +49,7 @@ const ArtistList = () => {
                 {
                     label: 'Artist ID',
                     field: 'id',
-                    sort: 'asc'
+                    sort: 'disabled'
                 },
                 {
                     label: 'Name',
@@ -53,7 +58,8 @@ const ArtistList = () => {
                 },  
                 {
                     label: 'Actions',
-                    field: 'actions'                  
+                    field: 'actions',
+                    sort: 'disabled'                  
                 }
             ],
             rows: []
@@ -62,18 +68,20 @@ const ArtistList = () => {
             data.rows.push({                
                 id: <small>{artist._id}</small>,
                 name: artist.name,
-                actions: <Fragment>
-                            <Link to={`/admin/artist/${artist._id}`}>
-                                <i className="fa fa-pencil"/>
-                            </Link> 
-                            &nbsp; &nbsp;                    
-                            <i 
-                                className="fa fa-trash-o"
-                                onClick={() => deleteCategoryHandler(artist._id)}
-                            />
-                        </Fragment> 
+                actions: 
+                    <Fragment>
+                        <Link to={`/admin/artist/${artist._id}`}>
+                            <i className="fa fa-pencil" />
+                        </Link> 
+                        &nbsp; &nbsp;                    
+                        <i 
+                            className="fa fa-trash-o"
+                            onClick={() => deleteCategoryHandler(artist._id)}
+                        />
+                    </Fragment> 
             })
         }) 
+
         return data
     }    
 
@@ -87,13 +95,17 @@ const ArtistList = () => {
 
                 <div className="wrapper parent dashboard">
 
-                    <aside><Sidebar /></aside>            
+                    <aside>
+                        
+                        <Sidebar />
+                        
+                    </aside>            
 
                     <article>                        
                         
                         {loading ? <Loader /> : (
 
-                            <div className="user-form cart mdb-table">
+                            <div className="user-form cart">
 
                                 <h1>Artist Category</h1>
 
@@ -103,7 +115,10 @@ const ArtistList = () => {
                                     </Link> 
                                 </p>                                                              
 
-                                <MDBDataTable data={setCategories()} />
+                                <MDBDataTable 
+                                    className="mdb-table"
+                                    data={setCategories()}                                     
+                                />
 
                                 <Link to="/dashboard"><i className="fa fa-times" /></Link>
