+++ frontend/src/components/admin/ProductsList.js
@@ -1,15 +1,14 @@
 import React, { Fragment, useEffect } from 'react'
-import { Link } from 'react-router-dom'
 import { MDBDataTable } from 'mdbreact'
-import MetaData from '../layouts/MetaData'
-import Loader from '../layouts/Loader'
-import Sidebar from '../admin/Sidebar'
 import { useAlert } from 'react-alert'
 import { useDispatch, useSelector } from 'react-redux'
-import { useNavigate } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { getAdminProducts, deleteProduct, clearErrors } from '../../actions/productActions'
 import { DELETE_PRODUCT_RESET } from '../../constants/productConstants'
 import FormattedPrice from '../layouts/FormattedPrice'
+import MetaData from '../layouts/MetaData'
+import Loader from '../layouts/Loader'
+import Sidebar from '../admin/Sidebar'
 
 const ProductsList = () => {
 
@@ -17,10 +16,11 @@ const ProductsList = () => {
     const dispatch = useDispatch()
     const navigate = useNavigate()
 
-    const { loading, error, products } = useSelector(state => state.products)
-    const { error: deleteError, isDeleted } = useSelector(state => state.product)
+    const { loading, error, products      } = useSelector( state => state.products )
+    const { error: deleteError, isDeleted } = useSelector( state => state.product  )
 
     useEffect(() => {
+
         dispatch(getAdminProducts())
 
         if(error) {
@@ -32,10 +32,10 @@ const ProductsList = () => {
             dispatch(clearErrors())
         }
         if(isDeleted) {
-            alert.success('Artwork deleted successfully')    
-            // navigate('/admin/products')        
+            alert.success('Artwork Deleted Successfully')    
             dispatch({ type: DELETE_PRODUCT_RESET })            
         }
+        
     }, [dispatch, navigate, alert, error, isDeleted, deleteError])
 
     const setProducts = () => {
@@ -44,11 +44,12 @@ const ProductsList = () => {
                 {
                     label: 'Preview',
                     field: 'url',
+                    sort: 'disabled'
                 },
                 {
                     label: 'Artwork ID',
                     field: 'id',
-                    sort: 'asc'
+                    sort: 'disabled'
                 },
                 {
                     label: 'Name',
@@ -58,7 +59,7 @@ const ProductsList = () => {
                 {
                     label: 'Price',
                     field: 'price',
-                    sort: 'asc'
+                    sort: 'disabled'
                 },
                 {
                     label: 'Stock',
@@ -67,7 +68,8 @@ const ProductsList = () => {
                 },
                 {
                     label: 'Actions',
-                    field: 'actions'                  
+                    field: 'actions',
+                    sort: 'disabled'                
                 }
             ],
             rows: []
@@ -88,26 +90,31 @@ const ProductsList = () => {
                 name: product.name,
                 price: <FormattedPrice number={product.price} />, 
                 stock: product.stock,
-                actions: <Fragment>
-                    <Link to={`/admin/product/${product._id}`}>
-                        <i className="fa fa-pencil"/>
-                    </Link> 
-                    &nbsp; &nbsp;                    
-                    <i 
-                        className="fa fa-trash-o"
-                        onClick={() => deleteProductHandler(product._id)}
-                    />
-                </Fragment> 
+                actions: 
+                    <Fragment>
+                        <Link to={`/admin/product/${product._id}`}>
+                            <i className="fa fa-pencil" />
+                        </Link> 
+                        &nbsp; &nbsp;                    
+                        <i 
+                            className="fa fa-trash-o"
+                            onClick={() => deleteProductHandler(product._id)}
+                        />
+                    </Fragment> 
             })
         })
+
         return data
     }
 
     const deleteProductHandler = (id) => {
-        dispatch(deleteProduct(id))
+        if ( window.confirm("Are you Sure?") === true ) {
+            dispatch(deleteProduct(id))
+        }       
     }
 
     return (
+
         <Fragment>
 
             <MetaData title={'All Products'} />
@@ -117,26 +124,27 @@ const ProductsList = () => {
                 <div className="wrapper parent dashboard">
 
                     <aside>
+
                         <Sidebar />
+
                     </aside>            
 
                     <article>
+
                         <Fragment>  
 
-                            <div className="user-form cart mdb-table">
+                            <div className="user-form cart">
 
                                 <h1>All Artwork</h1>
 
                                 {loading ? <Loader /> : (
-                                    <MDBDataTable                                    
-                                        data={setProducts()}
-                                        bordered
-                                        striped
-                                        hover    
+                                    <MDBDataTable  
+                                        className="mdb-table"                                  
+                                        data={setProducts()}                                
                                     />
                                 )}
 
-                            <Link to="/dashboard"><i className="fa fa-times"></i></Link>
+                                <Link to="/dashboard"><i className="fa fa-times" /></Link>
 
                             </div>
 
@@ -149,7 +157,9 @@ const ProductsList = () => {
             </div>
             
         </Fragment>
+
     )
+
 }
 
 export default ProductsList
