+++ frontend/src/components/cart/Cart.js
@@ -15,10 +15,14 @@ const Cart = () => {
     let totalPrice = cartItems.reduce((acc, item) => acc + item.quantity * item.price, 0)
 
     const removeCartItemHandler = (id) => {
-        dispatch(removeItemFromCart(id))
+        if ( window.confirm("Are you Sure?") === true ) {
+            dispatch(removeItemFromCart(id))
+        } 
     }
-    const emptyCartHandler = () => {        
-        dispatch(emptyCart())
+    const emptyCartHandler = () => {   
+        if ( window.confirm("Are you Sure?") === true ) {
+            dispatch(emptyCart())
+        }  
     }
     const increaseQty = (id, quantity, stock) => {
         const newQty = quantity + 1
